extract finding and creating users from controller
