+++ backend/app/controllers/user/getOrCreateUser.js
@@ -1,29 +1,12 @@
-const User = require( '../../../schema/user' );
-const DBConnection = require( '../../database/DBConnection' );
 const GoogleOAuthClient = require( '../../auth/GoogleOAuthClient' );
+const FindOrCreateUserService = require( '../../services/FindOrCreateUserService' );
 
 module.exports = async ( req, res ) => {
    try {
       const token = req.header( 'X-GOOGLE-AUTH' );
 
-      if ( !token ) {
-         throw new Error( 'Missing user token' );
-      }
-
-      const OAuthClient = new GoogleOAuthClient();
-      const googleUser = await OAuthClient.verify( token );
-
-      await new DBConnection().make();
-      let user = await User.findOne( { googleId: googleUser.id } ).exec();
-
-      if ( !user ) {
-         const newUser = {
-            name: googleUser.name,
-            email: googleUser.email,
-            googleId: googleUser.id,
-         };
-         user = await new User( newUser ).save();
-      }
+      const googleUser = await new GoogleOAuthClient().verify( token );
+      const user = await new FindOrCreateUserService().get( googleUser );
 
       res.send( user );
    }
